fix: empty SpecialSeparator deserializer crash + widen CModePacket.MorphSkin (#446)
* fix: empty SpecialSeparator on sub-packets, widen CModePacket.MorphSkin

**Deserializer — empty SpecialSeparator (non-breaking bug fix)**

Four packet declarations use [PacketIndex(n, SpecialSeparator = "")]
around an UpgradeRareSubPacket: EqPacket.WeaponUpgradeRarePacket /
ArmorUpgradeRarePacket, EquipPacket.WeaponUpgradeRareSubPacket /
ArmorUpgradeRareSubPacket, and InCharacterSubPacket.{Weapon,Armor}
UpgradeRareSubPacket. The serializer handles this correctly (joins
the sub-packet's two fields with no separator — e.g. Upgrade=7,
Rare=5 → "75" on the wire). The deserializer was doing
`matches[currentIndex].Replace("", " ")`, which .NET rejects with
ArgumentException. So every equip / eq / in packet with an
upgrade-rare field failed to parse.

Fix: branch on empty SpecialSeparator and insert a space between
every character via `string.Join(' ', chars)`. Existing separators
(".", "|", etc.) are unchanged.

**CModePacket.MorphSkin byte → short (BREAKING)**

Real captures consistently carry values of 7583/8151/8558/9471/etc.
in the last field, overflowing byte. Widening to short matches the
observed range. Source-level breaking because consumers pinning the
property type to byte stop compiling; runtime conversions widen safely.


* fix: inventory ivn sub-packet list non-nullable (clear slot as 0, not -1)

Real trace captures show an empty inventory slot as ivn {type} {slot}.0.0...
but NosCore was generating {slot}.-1.0... because the extension in NosCore
treats a null item-instance as a marker and hard-codes VNum = -1. The -1
sentinel is the Serializer's null-sub-packet output — which makes sense
for a nullable element, but the wire format never uses -1 here.

Tightening the packet contract: List<IvnSubPacket?>? -> List<IvnSubPacket>?
on IvnPacket, InvPacket, StashClientPacket, PStashClientPacket,
FStashClientPacket, FStashAllPacket. Now null entries are a compile error;
callers must produce a concrete sub-packet with VNum = 0 for cleared slots.
StashAllPacket and PStashAllPacket were already non-nullable — this aligns
the rest.

Regression test SerializeIvnPacketWithClearedSlotUsesZeroVNum asserts
"ivn 0 19.0.0.0.0" for a zeroed sub-packet so the wire output is pinned.

This unblocks the NosCore-side fix: ItemInstanceExtension.GenerateIvnSubPacket
returns VNum = 0 (not -1) when the item is null, so the client actually
clears the source slot after mvi / mve.


* fix: pad IvnSubPacket to 7 fields to match equipment wire format

Real trace captures of ivn with inventory type 0 (equipment) carry seven
dot-separated values — e.g. ivn 0 20.4998.0.0.0.0.0 — but the sub-packet
only declared five (Slot, VNum, RareAmount, UpgradeDesign, SecondUpgrade),
so NosCore was shipping slot.vnum.rare.upgrade.0 with two fields missing.
The client never cleared the source slot after an mvi because the packet
it received didn't match the expected shape.

Added Unknown5 (byte) and Unknown6 (short) at indices 5 and 6. Names are
placeholders — reverse-engineering suggests index 5 may be an IsBound
flag and index 6 the rune/option count — but without confirmation we
keep them opaque.

Tests pin both sides of a move to the exact trace:
- SerializeIvnPacketWithClearedSlotUsesZeroVNum → "ivn 0 19.0.0.0.0.0.0"
- SerializeIvnPacketWithPopulatedEquipmentSlotMatchesTrace → "ivn 0 20.4998.0.0.0.0.0"


* fix: rename audited Unknown fields and reorder InItemSubPacket to match trace

Audit against vanosilla (C:\Users\erwan\Desktop\vanosilla) named the
following fields that NosCore.Packets left as Unknown. Each rename is
backed by a direct vanosilla generator line at the same wire position.

Renames (all source-breaking, wire-identical):
- DropPacket.Unknown (idx 5, byte) -> IsQuest (bool).
  vanosilla MapExtensions.cs: $"drop {vnum} {id} {x} {y} {amount} {(isQuest ? 1 : 0)} {owner}"
- InCharacterSubPacket.Unknown (idx 11, byte) -> FairyBooster (bool).
  vanosilla CharacterPacketExtension.cs line 733: (IsUsingFairyBooster() ? 1 : 0)
- InCharacterSubPacket.Morph (idx 12) -> FairyMorph. Old name was wrong;
  vanosilla line 734 writes Fairy.GameItem.Morph here, not the character's.
- InCharacterSubPacket.Unknown2 (idx 13, byte) -> ShowInEffect (bool).
  vanosilla line 735: (showInEffect ? 0 : 1)
- InCharacterSubPacket.Unknown3 (idx 14, byte) -> Morph. This is the
  character's actual Morph (vanosilla line 736: character.Morph).
- ScnPacket.Unknown21 (idx 34, int) -> IsTeamMember (bool).
  vanosilla MatePacketExtensions.cs line 254: (mateEntity.IsTeamMember ? 1 : 0)
- ScnPacket.Unknown22 (idx 35, int) -> LevelXp. vanosilla line 255:
  _algorithm.GetLevelXp(mateEntity).

InItemSubPacket position fix
The current layout emits {amount} {isQuest} {owner} {0} but the real
trace for "in 9" and vanosilla both end {amount} {isQuest} 0 {owner}:
  trace: in 9 1046 6467573 52 154 10 0 0 14643732
The Unknown byte at index 3 was effectively the "0" that NosTale inserted
before ownerId in a protocol revision. NosCore put it at the wrong end.
Fix: move Unknown to index 2, Owner to index 3. Added regression test
SerializeInPacketForItemPutsOwnerAfterUnknownZero pinning the trace.

Not attempted (no vanosilla evidence at that index): ClistPacket,
NInvPacket, AtPacket, EquipmentSubPacket, PinitSubPacket, MlInfoPacket,
BfePacket, ScPStcPacket, EffObPacket, EffTPacket, RsfiPacket, CtlPacket,
StpSubPacket, LevPacket, ScnPacket.Unknown/Unknown2 (vanosilla hardcodes 0),
CInfoPacket.Unknown1 (vanosilla hardcodes a dash), IvnSubPacket.Unknown5/6
(vanosilla knows RunesCount but trace has one extra unidentified field,
so the position of RunesCount is ambiguous without a runes-populated
capture).


* chore: bump NosCore.Packets to 17.0.0

Major because this PR now carries several source-breaking changes beyond
the original CModePacket.MorphSkin widening: IvnPacket/InvPacket/Stash*
move to List<IvnSubPacket> (non-nullable items), InItemSubPacket swaps
indices 2 and 3 so Owner sits after the Unknown zero, and the audited
rename pass touches DropPacket / InCharacterSubPacket / ScnPacket.

---------

+++ src/NosCore.Packets/ServerPackets/Player/CModePacket.cs
@@ -35,6 +35,6 @@ namespace NosCore.Packets.ServerPackets.Player
         public byte Size { get; set; }
 
         [PacketIndex(7)]
-        public byte MorphSkin { get; set; }
+        public short MorphSkin { get; set; }
     }
 }
